+++ src/conf/conf.cpp
@@ -41,6 +41,7 @@ namespace conf {
     std::vector<std::unique_ptr<Match>> matchConfigs;
     std::vector<std::string> INDEX_FILES;
     std::vector<std::unique_ptr<Redirect>> redirectRules;
+    std::vector<std::unique_ptr<Rewrite>> rewriteRules;
 
     std::filesystem::path ACCESS_LOG_FILE;
     std::filesystem::path ERROR_LOG_FILE;
@@ -248,7 +249,7 @@ namespace conf {
                 return CONF_FAILURE;
         #endif
 
-        /************ LOAD MATCHES, REDIRECTS, & MIMES ************/
+        /************ LOAD MATCHES, REDIRECTS/REWRITES, & MIMES ************/
 
         pugi::xml_object_range matchNodes = root.children("Match");
         for (pugi::xml_node& match : matchNodes) {
@@ -264,6 +265,14 @@ namespace conf {
             std::unique_ptr<Redirect> pRedirect = loadRedirect(redirectRule);
             if (pRedirect == nullptr) return CONF_FAILURE;
             redirectRules.push_back( std::move(pRedirect) );
+        }
+
+        pugi::xml_object_range rewriteRuleNodes = root.children("Rewrite");
+        for (pugi::xml_node& rewriteRule : rewriteRuleNodes) {
+            // Parse match
+            std::unique_ptr<Rewrite> pRewrite = loadRewrite(rewriteRule);
+            if (pRewrite == nullptr) return CONF_FAILURE;
+            rewriteRules.push_back( std::move(pRewrite) );
         }
 
         if (loadMIMES() == CONF_FAILURE)