+++ src/http/request.hpp
@@ -1,6 +1,8 @@
 #ifndef __HTTP_REQUEST_HPP
 #define __HTTP_REQUEST_HPP
 
+#include <optional>
+
 #include "../pch/common.hpp"
 
 #include "http_tools.hpp"
@@ -8,26 +10,11 @@
 
 namespace http {
 
-    enum METHOD {
-        GET = 0,
-        HEAD = 1,
-        OPTIONS = 2,
-        POST = 3,
-        PUT = 4,
-        DEL = 5,
-        PATCH = 6,
-        UNKNOWN = 999
-    };
-
-    // Used to extract headers and status info from partial request
-    typedef std::unordered_map<std::string, std::string> headers_map_t;
-    void loadEarlyHeaders(headers_map_t&, const std::string&);
-
     class Request {
         public:
             Request(headers_map_t& headers, const std::string&, std::string, const bool, const bool);
 
-            const std::string* getHeader(std::string) const;
+            std::optional<std::string> getHeader(std::string) const;
             inline const std::string getIPStr() const { return ipStr; };
             inline METHOD getMethod() const { return method; };
             inline const std::string& getMethodStr() const { return methodStr; };
@@ -35,7 +22,9 @@ namespace http {
             inline const std::string& getPathStr() const { return pathStr; };
             inline const std::string& getBody() const { return body; };
             inline const std::string& getVersion() const { return httpVersionStr; };
-            inline int getCompressMethod() const { return compressMethod; };
+            int getCompressMethod(const std::string& MIME) const;
+
+            void rewriteRawPath(const std::string& newPath);
 
             inline const headers_map_t& getHeaders() const { return headers; };
             inline const std::vector<std::pair<size_t, size_t>>& getByteRanges() const { return byteRanges; };
@@ -73,7 +62,7 @@ namespace http {
             std::string httpVersionStr;
 
             std::string body;
-            int compressMethod = NO_COMPRESS;
+            int compressMethods = NO_COMPRESS;
     };
 
 }