+++ src/http/version/handler_1_0.cpp
@@ -1,5 +1,7 @@
 #include "handler_1_0.hpp"
 
+#include <optional>
+
 #include "../../conf/conf.hpp"
 #include "../../util/toolbox.hpp"
 #include "../../logs/logger.hpp"
@@ -26,6 +28,30 @@ namespace http::version::handler_1_0 {
         }
 
         // Verify no 400 errors have been met
+        if ( request.has400Error() ) {
+            setStatusMaybeErrorDoc(request, *pResponse, 400);
+            return pResponse;
+        }
+
+        // Handle rewrites
+        if (!conf::rewriteRules.empty()) {
+            // Format raw path & break off query string
+            std::string rawPath = request.getRawPathStr();
+            const size_t queryIndex = rawPath.find('?');
+            const std::string queryString = queryIndex != std::string::npos ? rawPath.substr(queryIndex) : "";
+            while (queryIndex != std::string::npos && rawPath.size() > queryIndex)
+                rawPath.pop_back();
+
+            // Decode URI (no need to try-catch, done in constructor)
+            decodeURI(rawPath);
+
+            // Check rewrite rule patterns
+            for (const std::unique_ptr<conf::Rewrite>& pRewrite : conf::rewriteRules)
+                if (pRewrite->loadRewrittenPath(rawPath))
+                    request.rewriteRawPath(rawPath);
+        }
+
+        // Check once more for a 400 error after parsing rewrite rules
         if ( request.has400Error() ) {
             setStatusMaybeErrorDoc(request, *pResponse, 400);
             return pResponse;
@@ -65,15 +91,18 @@ namespace http::version::handler_1_0 {
             setStatusMaybeErrorDoc(request, *pResponse, 413);
             return pResponse;
         }
-
+        
         // Handle redirects
         if (!conf::redirectRules.empty()) {
             // Format raw path & break off query string
-            std::string rawPath = request.getPathStr();
+            std::string rawPath = request.getRawPathStr();
             const size_t queryIndex = rawPath.find('?');
             const std::string queryString = queryIndex != std::string::npos ? rawPath.substr(queryIndex) : "";
             while (queryIndex != std::string::npos && rawPath.size() > queryIndex)
                 rawPath.pop_back();
+
+            // Decode URI
+            decodeURI(rawPath);
 
             // Check redirect rule patterns
             std::string locationBuf;
@@ -85,7 +114,7 @@ namespace http::version::handler_1_0 {
                 const unsigned int status = pRedirect->getStatus();
                 if (status > 302) ERROR_LOG << "HTTP/1.0 falling back from " << status << " status to 302 status" << std::endl;
                 pResponse->setStatus( status > 302 ? 302 : status );
-                pResponse->setHeader( "Location", locationBuf );
+                pResponse->setHeader( "Location", locationBuf + queryString );
                 return pResponse;
             }
         }
@@ -111,8 +140,7 @@ namespace http::version::handler_1_0 {
 
                 // Check if previously cached
                 const auto pLastModTS = request.getHeader("IF-MODIFIED-SINCE");
-                if (pLastModTS != nullptr) {
-                    // Compare timestamps
+                if (pLastModTS.has_value()) { // Compare timestamps
                     try {
                         // serverTime <= clientTime
                         if (getFileModTimeT(file.path) <= getTimeTFromGMT(*pLastModTS)) {