+++ src/http/version/handler_1_1.cpp
@@ -1,5 +1,7 @@
 #include "handler_1_1.hpp"
 
+#include <optional>
+
 #include "../cgi/process.hpp"
 #include "../../conf/conf.hpp"
 #include "../../util/toolbox.hpp"
@@ -32,6 +34,30 @@ namespace http::version::handler_1_1 {
         // Verify no 400 errors have been met
         if ( request.has400Error() ) {
             pResponse->loadBodyFromErrorDoc(400);
+            return pResponse;
+        }
+
+        // Handle rewrites
+        if (!conf::rewriteRules.empty()) {
+            // Format raw path & break off query string
+            std::string rawPath = request.getRawPathStr();
+            const size_t queryIndex = rawPath.find('?');
+            const std::string queryString = queryIndex != std::string::npos ? rawPath.substr(queryIndex) : "";
+            while (queryIndex != std::string::npos && rawPath.size() > queryIndex)
+                rawPath.pop_back();
+
+            // Decode URI (no need to try-catch, done in constructor)
+            decodeURI(rawPath);
+
+            // Check rewrite rule patterns
+            for (const std::unique_ptr<conf::Rewrite>& pRewrite : conf::rewriteRules)
+                if (pRewrite->loadRewrittenPath(rawPath))
+                    request.rewriteRawPath(rawPath);
+        }
+
+        // Check once more for a 400 error after parsing rewrite rules
+        if ( request.has400Error() ) {
+            setStatusMaybeErrorDoc(request, *pResponse, 400);
             return pResponse;
         }
 
@@ -73,11 +99,14 @@ namespace http::version::handler_1_1 {
         // Handle redirects
         if (!conf::redirectRules.empty()) {
             // Format raw path & break off query string
-            std::string rawPath = request.getPathStr();
+            std::string rawPath = request.getRawPathStr();
             const size_t queryIndex = rawPath.find('?');
             const std::string queryString = queryIndex != std::string::npos ? rawPath.substr(queryIndex) : "";
             while (queryIndex != std::string::npos && rawPath.size() > queryIndex)
                 rawPath.pop_back();
+
+            // Decode URI
+            decodeURI(rawPath);
 
             // Check redirect rule patterns
             std::string locationBuf;
@@ -87,7 +116,7 @@ namespace http::version::handler_1_1 {
 
                 // New location found, set status
                 pResponse->setStatus( pRedirect->getStatus() );
-                pResponse->setHeader( "Location", locationBuf );
+                pResponse->setHeader( "Location", locationBuf + queryString );
                 return pResponse;
             }
         }
@@ -135,7 +164,7 @@ namespace http::version::handler_1_1 {
 
                 // Check if previously cached
                 const auto pLastModTS = request.getHeader("IF-MODIFIED-SINCE");
-                if (pLastModTS != nullptr) {
+                if (pLastModTS.has_value()) {
                     // Compare timestamps
                     try {
                         // serverTime <= clientTime